Tests failing due to openscm append weakness
