+++ pymagicc/scenarios/__init__.py
@@ -2,47 +2,12 @@ from os.path import dirname, join, abspath
 from copy import deepcopy
 
 
-from ..io import MAGICCData
+from ..io import MAGICCData, read_scen_file
 from ..config import default_config
 
 
 # path to load files from included package
 _magicc6_included_distribution_path = dirname(default_config["EXECUTABLE_6"])
-
-
-def read_scen_file(
-    filepath,
-    columns={
-        "model": ["unspecified"],
-        "scenario": ["unspecified"],
-        "climate_model": ["unspecified"],
-    },
-    **kwargs
-):
-    """
-    Read a MAGICC .SCEN file.
-
-    Parameters
-    ----------
-    filepath : str
-        Filepath of the .SCEN file to read
-
-    columns : dict
-        Passed to ``__init__`` method of MAGICCData. See 
-        ``MAGICCData.__init__`` for details.
-
-    kwargs
-        Passed to ``__init__`` method of MAGICCData. See 
-        ``MAGICCData.__init__`` for details.
-
-    Returns
-    -------
-    :obj:`pymagicc.io.MAGICCData`
-        ``MAGICCData`` object containing the data and metadata.
-    """
-    mdata = MAGICCData(filepath, columns=columns, **kwargs)
-
-    return mdata
 
 
 rcp26 = read_scen_file(
